Rename edit command to copy
